wm: set SDL_VIDEO_X11_LEGACY_FULLSCREEN implicitly
SDL2 falls back on a legacy fullscreen method that breaks KivyWM on
startup if it doesn't detect that the WM supports the _NET_WM protocol [1].

Ideally, we would create the WM's window ourself, apply the appropriate
properties, and have Kivy use it for its application window. This is a
hack to make it work as it is.

[1] https://forums.libsdl.org/viewtopic.php?p=34362