windowmanager: fix refresh with drivers other than i965
The previous redraw method exploited an implementation detail of Mesa,
and the i965 graphics driver to avoid unbinding and rebinding the
texture every frame. This does not work with many other graphics
drivers.

This commit creates a new unbinds the texture, creates a new pixmap, and
rebinds the texture every frame.